refactor(ai): remove provider transport package cycles
