+++ packages/ai/src/providers/openai-completions.ts
@@ -3,13 +3,10 @@ import OpenAI from "openai";
 import type {
   ChatCompletionAssistantMessageParam,
   ChatCompletionChunk,
-  ChatCompletionContentPart,
-  ChatCompletionContentPartImage,
   ChatCompletionContentPartText,
   ChatCompletionDeveloperMessageParam,
   ChatCompletionMessageParam,
   ChatCompletionSystemMessageParam,
-  ChatCompletionToolMessageParam,
 } from "openai/resources/chat/completions.js";
 import { getEnvApiKey } from "../env-api-keys.js";
 import { getAiTransportHost } from "../host.js";
@@ -18,6 +15,8 @@ import {
   calculateCost,
   clampThinkingLevel,
 } from "../model-utils.js";
+import { convertMessages } from "../openai-completions-messages.js";
+import type { OpenAICompletionsOptions } from "../provider-options.js";
 import {
   resolveOpenAICompletionsCompat,
   type ResolvedOpenAICompletionsCompat,
@@ -32,7 +31,6 @@ import type {
   Model,
   SimpleStreamOptions,
   StreamFunction,
-  StreamOptions,
   TextContent,
   ThinkingContent,
   Tool,
@@ -49,17 +47,13 @@ import { headersToRecord } from "../utils/headers.js";
 import { parseStreamingJson } from "../utils/json-parse.js";
 import { formatProviderError } from "../utils/provider-error.js";
 import { createReasoningTagTextPartitioner } from "../utils/reasoning-tag-text-partitioner.js";
-import { sanitizeSurrogates } from "../utils/sanitize-unicode.js";
 import {
   createFirstStreamEventAbortController,
   getFirstStreamEventTimeoutHandler,
   getFirstStreamEventTimeoutMs,
   withFirstStreamEventTimeout,
 } from "../utils/stream-first-event-timeout.js";
-import {
-  splitSystemPromptCacheBoundary,
-  stripSystemPromptCacheBoundary,
-} from "../utils/system-prompt-cache-boundary.js";
+import { splitSystemPromptCacheBoundary } from "../utils/system-prompt-cache-boundary.js";
 import { resolveCacheRetention } from "./cache-retention.js";
 import { isCloudflareProvider, resolveCloudflareBaseUrl } from "./cloudflare.js";
 import { buildCopilotDynamicHeaders, hasCopilotVisionInput } from "./github-copilot-headers.js";
@@ -72,16 +66,9 @@ import { mapOpenAIStopReason } from "./openai-stop-reason.js";
 import {
   projectOpenAITools,
   reconcileOpenAICompletionsToolChoice,
-  type OpenAICompletionsToolChoice,
   type OpenAIToolProjection,
 } from "./openai-tool-projection.js";
 import { buildBaseOptions } from "./simple-options.js";
-import {
-  describeToolResultMediaPlaceholder,
-  extractToolResultText,
-  isImageWithMediaPayload,
-} from "./tool-result-text.js";
-import { transformMessages } from "./transform-messages.js";
 
 /**
  * Check if conversation messages contain tool calls or tool results.
@@ -104,29 +91,8 @@ function hasToolHistory(messages: Message[]): boolean {
   return false;
 }
 
-function isTextContentBlock(block: { type: string }): block is TextContent {
-  return block.type === "text";
-}
-
-function isThinkingContentBlock(block: { type: string }): block is ThinkingContent {
-  return block.type === "thinking";
-}
-
-function isToolCallBlock(block: { type: string }): block is ToolCall {
-  return block.type === "toolCall";
-}
-
-const EMPTY_TOOL_RESULT_TEXT = "(no output)";
-
-function sanitizeToolResultText(text: string, fallback: string): string {
-  const sanitized = sanitizeSurrogates(text);
-  return sanitized.trim().length > 0 ? sanitized : fallback;
-}
-
-export interface OpenAICompletionsOptions extends StreamOptions {
-  toolChoice?: OpenAICompletionsToolChoice;
-  reasoningEffort?: "minimal" | "low" | "medium" | "high" | "xhigh";
-}
+export type { OpenAICompletionsOptions } from "../provider-options.js";
+export { convertMessages } from "../openai-completions-messages.js";
 
 interface OpenAICompatCacheControl {
   type: "ephemeral";
@@ -842,13 +808,16 @@ function buildParams(
   // Provider compat is authoritative; keep model-level and literal values as fallbacks
   // for catalogs that have not adopted reasoningEffortMap.
   const reasoningEffortMap = resolveOpenAIReasoningEffortMap(model);
+  const thinkingLevelMap = model.thinkingLevelMap as
+    | Partial<Record<NonNullable<OpenAICompletionsOptions["reasoningEffort"]>, string | null>>
+    | undefined;
   const reasoningEffort =
     options?.reasoningEffort === undefined
       ? undefined
       : (reasoningEffortMap[options.reasoningEffort] ??
-        model.thinkingLevelMap?.[options.reasoningEffort] ??
+        thinkingLevelMap?.[options.reasoningEffort] ??
         options.reasoningEffort);
-  const reasoningEnabled = reasoningEffort !== undefined;
+  const reasoningEnabled = reasoningEffort !== undefined && reasoningEffort !== "none";
   const offReasoningEffort = reasoningEffortMap.off ?? model.thinkingLevelMap?.off;
 
   if (compat.thinkingFormat === "zai" && model.reasoning) {
@@ -1067,295 +1036,6 @@ function buildCacheControlledTextParts(
     parts.push({ type: "text", text: split.dynamicSuffix });
   }
   return parts.length > 0 ? parts : [{ type: "text", text: "" }];
-}
-
-export function convertMessages(
-  model: Model<"openai-completions">,
-  context: Context,
-  compat: ResolvedOpenAICompletionsCompat,
-  options: {
-    cacheOptOutIndexes?: Set<number>;
-    preserveSystemPromptCacheBoundary?: boolean;
-  } = {},
-): ChatCompletionMessageParam[] {
-  const params: ChatCompletionMessageParam[] = [];
-
-  const normalizeToolCallId = (id: string): string => {
-    // Handle pipe-separated IDs from OpenAI Responses API
-    // Format: {call_id}|{id} where {id} can be 400+ chars with special chars (+, /, =)
-    // These come from providers like github-copilot, openai, opencode
-    // Extract just the call_id part and normalize it
-    if (id.includes("|")) {
-      const callId = id.slice(0, id.indexOf("|"));
-      // Sanitize to allowed chars and truncate to 40 chars (OpenAI limit)
-      return callId.replace(/[^a-zA-Z0-9_-]/g, "_").slice(0, 40);
-    }
-
-    if (model.provider === "openai") {
-      return id.length > 40 ? id.slice(0, 40) : id;
-    }
-    return id;
-  };
-
-  const transformedMessages = transformMessages(context.messages, model, (id) =>
-    normalizeToolCallId(id),
-  );
-
-  if (context.systemPrompt) {
-    const useDeveloperRole = model.reasoning && compat.supportsDeveloperRole;
-    const role = useDeveloperRole ? "developer" : "system";
-    const systemPrompt = options.preserveSystemPromptCacheBoundary
-      ? context.systemPrompt
-      : stripSystemPromptCacheBoundary(context.systemPrompt);
-    params.push({
-      role,
-      content: sanitizeSurrogates(systemPrompt),
-    });
-  }
-
-  let lastRole: string | null = null;
-
-  for (let i = 0; i < transformedMessages.length; i++) {
-    const msg = transformedMessages[i];
-    if (!msg) {
-      continue;
-    }
-    // Some providers don't allow user messages directly after tool results
-    // Insert a synthetic assistant message to bridge the gap
-    if (
-      compat.requiresAssistantAfterToolResult &&
-      lastRole === "toolResult" &&
-      msg.role === "user"
-    ) {
-      params.push({
-        role: "assistant",
-        content: "I have processed the tool results.",
-      });
-    }
-
-    if (msg.role === "user") {
-      const isRuntimeContextCarrier = msg.runtimeContextCarrier === true;
-      if (typeof msg.content === "string") {
-        const userParam: ChatCompletionMessageParam = {
-          role: "user",
-          content: sanitizeSurrogates(msg.content),
-        };
-        if (isRuntimeContextCarrier) {
-          options.cacheOptOutIndexes?.add(params.length);
-        }
-        params.push(userParam);
-      } else {
-        const content: ChatCompletionContentPart[] = msg.content.map(
-          (item): ChatCompletionContentPart => {
-            if (item.type === "text") {
-              return {
-                type: "text",
-                text: sanitizeSurrogates(item.text),
-              } satisfies ChatCompletionContentPartText;
-            }
-            return {
-              type: "image_url",
-              image_url: {
-                url: `data:${item.mimeType};base64,${item.data}`,
-              },
-            } satisfies ChatCompletionContentPartImage;
-          },
-        );
-        if (content.length === 0) {
-          continue;
-        }
-        const userParam: ChatCompletionMessageParam = {
-          role: "user",
-          content,
-        };
-        if (isRuntimeContextCarrier) {
-          options.cacheOptOutIndexes?.add(params.length);
-        }
-        params.push(userParam);
-      }
-    } else if (msg.role === "assistant") {
-      // Some providers don't accept null content, use empty string instead
-      const assistantMsg: ChatCompletionAssistantMessageParam = {
-        role: "assistant",
-        content: compat.requiresAssistantAfterToolResult ? "" : null,
-      };
-
-      const assistantTextParts = msg.content
-        .filter(isTextContentBlock)
-        .filter((block) => block.text.trim().length > 0)
-        .map(
-          (block) =>
-            ({
-              type: "text",
-              text: sanitizeSurrogates(block.text),
-            }) satisfies ChatCompletionContentPartText,
-        );
-      const assistantText = assistantTextParts.map((part) => part.text).join("");
-
-      const nonEmptyThinkingBlocks = msg.content
-        .filter(isThinkingContentBlock)
-        .filter((block) => block.thinking.trim().length > 0);
-      if (nonEmptyThinkingBlocks.length > 0) {
-        if (compat.requiresThinkingAsText) {
-          // Convert thinking blocks to plain text (no tags to avoid model mimicking them)
-          const thinkingText = nonEmptyThinkingBlocks
-            .map((block) => sanitizeSurrogates(block.thinking))
-            .join("\n\n");
-          assistantMsg.content = [{ type: "text", text: thinkingText }, ...assistantTextParts];
-        } else {
-          // Always send assistant content as a plain string (OpenAI Chat Completions
-          // API standard format). Sending as an array of {type:"text", text:"..."}
-          // objects is non-standard and causes some models (e.g. DeepSeek V3.2 via
-          // NVIDIA NIM) to mirror the content-block structure literally in their
-          // output, producing recursive nesting like [{'type':'text','text':'[{...}]'}].
-          if (assistantText.length > 0) {
-            assistantMsg.content = assistantText;
-          }
-
-          // Use the signature from the first thinking block if available (for llama.cpp server + gpt-oss)
-          let signature = nonEmptyThinkingBlocks.at(0)?.thinkingSignature;
-          if (model.provider === "opencode-go" && signature === "reasoning") {
-            signature = "reasoning_content";
-          }
-          if (signature && signature.length > 0) {
-            (assistantMsg as typeof assistantMsg & Record<string, unknown>)[signature] =
-              nonEmptyThinkingBlocks.map((block) => block.thinking).join("\n");
-          }
-        }
-      } else if (assistantText.length > 0) {
-        // Always send assistant content as a plain string (OpenAI Chat Completions
-        // API standard format). Sending as an array of {type:"text", text:"..."}
-        // objects is non-standard and causes some models (e.g. DeepSeek V3.2 via
-        // NVIDIA NIM) to mirror the content-block structure literally in their
-        // output, producing recursive nesting like [{'type':'text','text':'[{...}]'}].
-        assistantMsg.content = assistantText;
-      }
-
-      const toolCalls = msg.content.filter(isToolCallBlock);
-      if (toolCalls.length > 0) {
-        assistantMsg.tool_calls = toolCalls.map((tc) => ({
-          id: tc.id,
-          type: "function" as const,
-          function: {
-            name: tc.name,
-            arguments: JSON.stringify(tc.arguments),
-          },
-        }));
-        const reasoningDetails = toolCalls.flatMap((tc) => {
-          const signature = tc.thoughtSignature;
-          if (!signature) {
-            return [];
-          }
-          try {
-            const parsed: unknown = JSON.parse(signature);
-            return parsed ? [parsed] : [];
-          } catch {
-            return [];
-          }
-        });
-        if (reasoningDetails.length > 0) {
-          (
-            assistantMsg as typeof assistantMsg & { reasoning_details?: unknown }
-          ).reasoning_details = reasoningDetails;
-        }
-      }
-      if (
-        compat.requiresReasoningContentOnAssistantMessages &&
-        model.reasoning &&
-        (assistantMsg as { reasoning_content?: string }).reasoning_content === undefined
-      ) {
-        (assistantMsg as { reasoning_content?: string }).reasoning_content = "";
-      }
-      // Skip assistant messages that have no content and no tool calls.
-      // Some providers require "either content or tool_calls, but not none".
-      // Other providers also don't accept empty assistant messages.
-      // This handles aborted assistant responses that got no content.
-      const content = assistantMsg.content;
-      const hasContent =
-        content !== null &&
-        content !== undefined &&
-        (typeof content === "string" ? content.length > 0 : content.length > 0);
-      if (!hasContent && !assistantMsg.tool_calls) {
-        continue;
-      }
-      params.push(assistantMsg);
-    } else if (msg.role === "toolResult") {
-      const imageBlocks: Array<{ type: "image_url"; image_url: { url: string } }> = [];
-      let j = i;
-
-      while (j < transformedMessages.length) {
-        const toolMsg = transformedMessages.at(j);
-        if (toolMsg?.role !== "toolResult") {
-          break;
-        }
-
-        // Extract text and image content
-        const textResult = extractToolResultText(toolMsg.content);
-        const mediaPlaceholder = describeToolResultMediaPlaceholder(toolMsg.content);
-        const hasImages = toolMsg.content.some(isImageWithMediaPayload);
-
-        // Always send tool result with text (or placeholder if only images)
-        const content = sanitizeToolResultText(
-          textResult,
-          mediaPlaceholder ?? EMPTY_TOOL_RESULT_TEXT,
-        );
-        // Some providers require the 'name' field in tool results
-        const toolResultMsg: ChatCompletionToolMessageParam = {
-          role: "tool",
-          content,
-          tool_call_id: toolMsg.toolCallId,
-        };
-        if (compat.requiresToolResultName && toolMsg.toolName) {
-          (toolResultMsg as typeof toolResultMsg & { name?: string }).name = toolMsg.toolName;
-        }
-        params.push(toolResultMsg);
-
-        if (hasImages && model.input.includes("image")) {
-          for (const block of toolMsg.content) {
-            if (isImageWithMediaPayload(block)) {
-              imageBlocks.push({
-                type: "image_url",
-                image_url: {
-                  url: `data:${block.mimeType};base64,${block.data}`,
-                },
-              });
-            }
-          }
-        }
-        j += 1;
-      }
-
-      i = j - 1;
-
-      if (imageBlocks.length > 0) {
-        if (compat.requiresAssistantAfterToolResult) {
-          params.push({
-            role: "assistant",
-            content: "I have processed the tool results.",
-          });
-        }
-
-        params.push({
-          role: "user",
-          content: [
-            {
-              type: "text",
-              text: "Attached image(s) from tool result:",
-            },
-            ...imageBlocks,
-          ],
-        });
-        lastRole = "user";
-      } else {
-        lastRole = "toolResult";
-      }
-      continue;
-    }
-
-    lastRole = msg.role;
-  }
-
-  return params;
 }
 
 function convertTools(
