+++ packages/ai/src/transports/anthropic-payload-policy.ts
@@ -1,7 +1,7 @@
 import {
   splitSystemPromptCacheBoundary,
   stripSystemPromptCacheBoundary,
-} from "../internal/shared.js";
+} from "../utils/system-prompt-cache-boundary.js";
 /**
  * Anthropic-family request payload policy helpers.
  * Applies service-tier and cache-control markers only when provider endpoint
