+++ packages/ai/src/transports/openai-completions-transport.ts
@@ -4,34 +4,27 @@ import { isRecord } from "@openclaw/normalization-core/record-coerce";
 import { uniqueStrings } from "@openclaw/normalization-core/string-normalization";
 import OpenAI from "openai";
 import type { ChatCompletionChunk } from "openai/resources/chat/completions.js";
+import { getEnvApiKey } from "../env-api-keys.js";
+import { applyProviderReportedUsageCost, calculateCost } from "../model-utils.js";
+import { convertMessages } from "../openai-completions-messages.js";
+import type { OpenAICompletionsOptions } from "../provider-options.js";
 import {
-  convertMessages,
   isOpenAIGpt54MiniModel,
   isOpenAIGpt55Model,
   isOpenAIGpt56Model,
-  mapOpenAIStopReason,
-  normalizeOpenAIStrictToolParameters,
-  projectOpenAITools,
-  reconcileOpenAICompletionsToolChoice,
   resolveOpenAIReasoningEffortForModel,
   type OpenAIReasoningEffort,
-} from "../internal/openai.js";
-import {
-  applyProviderReportedUsageCost,
-  calculateCost,
-  createFirstStreamEventAbortController,
-  createReasoningTagTextPartitioner,
-  getEnvApiKey,
-  getFirstStreamEventTimeoutHandler,
-  getFirstStreamEventTimeoutMs,
-  parseStreamingJson,
-  withFirstStreamEventTimeout,
-} from "../internal/runtime.js";
-import { stripSystemPromptCacheBoundary } from "../internal/shared.js";
+} from "../providers/openai-reasoning-effort.js";
 import {
   resolveOpenAICompletionsResponseFormat,
   shouldOmitOllamaCompatResponseFormat,
 } from "../providers/openai-response-format.js";
+import { mapOpenAIStopReason } from "../providers/openai-stop-reason.js";
+import {
+  projectOpenAITools,
+  reconcileOpenAICompletionsToolChoice,
+} from "../providers/openai-tool-projection.js";
+import { normalizeOpenAIStrictToolParameters } from "../providers/openai-tool-schema.js";
 import {
   clearPendingCommentaryText,
   rememberPendingCommentaryTags,
@@ -39,6 +32,15 @@ import {
   type PendingCommentaryTags,
 } from "../utils/assistant-text-phase.js";
 import { createAssistantMessageEventStream } from "../utils/event-stream.js";
+import { parseStreamingJson } from "../utils/json-parse.js";
+import { createReasoningTagTextPartitioner } from "../utils/reasoning-tag-text-partitioner.js";
+import {
+  createFirstStreamEventAbortController,
+  getFirstStreamEventTimeoutHandler,
+  getFirstStreamEventTimeoutMs,
+  withFirstStreamEventTimeout,
+} from "../utils/stream-first-event-timeout.js";
+import { stripSystemPromptCacheBoundary } from "../utils/system-prompt-cache-boundary.js";
 import { createDeepSeekTextFilter } from "./deepseek-text-filter.js";
 import {
   buildGuardedModelFetch,
@@ -71,7 +73,6 @@ import {
   sortTransportToolsByName,
   throwIfModelStreamAborted,
   type MutableAssistantOutput,
-  type OpenAICompletionsOptions,
   type OpenAIModeModel,
 } from "./openai-transport-shared.js";
 import { failTransportStream, finalizeTransportStream } from "./transport-stream-shared.js";
